Re-org files, update doc
