+++ gal.go
@@ -1,7 +1,5 @@
 package gal
 
-import "fmt"
-
 type exprType int
 
 const (
@@ -18,31 +16,6 @@ const (
 	objectAccessorByPropertyType // represents an object accessor of a "left hand side" expression by property
 	objectAccessorByMethodType   // represents an object accessor of a "left hand side" expression by method
 )
-
-type Value interface {
-	// Calculation
-	Add(Value) Value
-	Sub(Value) Value
-	Multiply(Value) Value
-	Divide(Value) Value
-	PowerOf(Value) Value
-	Mod(Value) Value
-	LShift(Value) Value
-	RShift(Value) Value
-	// Logical
-	LessThan(Value) Bool
-	LessThanOrEqual(Value) Bool
-	EqualTo(Value) Bool
-	NotEqualTo(Value) Bool
-	GreaterThan(Value) Bool
-	GreaterThanOrEqual(Value) Bool
-	And(Value) Bool
-	Or(Value) Bool
-	// Helpers
-	Stringer
-	fmt.Stringer
-	entry
-}
 
 // Example: Parse("blah").Eval(WithVariables(...), WithFunctions(...), WithObjects(...))
 // This allows to parse an expression and then use the resulting Tree for multiple
